requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
scikit-learn==1.4.1
google-generativeai==0.3.2
python-dotenv==1.0.1
folium==0.15.1
seaborn==0.13.2
matplotlib==3.8.3
streamlit-option-menu==0.3.12
streamlit-extras==0.3.1
geopy==2.4.1 